% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xache.R
\name{xache}
\alias{xache}
\title{xache}
\usage{
xache(symbol, file, init, ..., verbose = FALSE, force = FALSE,
  envir = parent.frame())
}
\arguments{
\item{symbol}{A string representing the variable that holds the results.}

\item{file}{A path towards a RData file from which to load and to which to #'   save the contents of the above variable.}

\item{init}{A function that will generate the contents if it does not already exist.}

\item{...}{Arguments to pass to the init function.}

\item{verbose}{Whether to print debug messages to screen.  Defaults to FALSE.}

\item{force}{Whether to force recomputation.  Defaults to FALSE.}

\item{envir}{The environment in which to look for, and if it does not already exist, to create the variable of interest.}
}
\description{
`xache` works as follows:
}
\details{
1. If the specified variable exists, nothing happens because there is no need
   to recompute its contents.
2. If it the variable does not exist, `xache` will attempt to load it from the
   specified RData file.
3. If the RData file does not exist, the specified function will evaluate the
   specified function, and store its results in the specified variable as well
   as in the RData file.
}
\examples{
xache("test", "test.RData", function () {
 # do some calculation
 2 * 2 # dummy result
})
}

